move fee hook impl
