+++ lib/el_golang_lib/hook/fee_hook.go
@@ -1,4 +1,4 @@
-package rollup
+package hook
 
 import (
 	"bytes"
@@ -34,13 +34,25 @@ type RollupConfig interface {
 func MakeSpecularEVMPreTransferHook(cfg RollupConfig) vm.EVMHook {
 	log.Info("Injected Specular EVM hook")
 	log.Info("L1Oracle config", "address", cfg.GetL1OracleAddress(), "baseFeeSlot", cfg.GetL1OracleBaseFeeSlot())
-	return func(msg types.Message, evm *vm.EVM) error {
+	return func(msg types.Message, db vm.StateDB) error {
 		tx := transactionFromMessage(msg, cfg)
-		fee, err := calculateL1Fee(tx, evm, cfg)
+		fee, err := calculateL1Fee(tx, db, cfg)
 		if err != nil {
 			return err
 		}
-		return chargeL1Fee(fee, msg, evm, cfg)
+
+		return chargeL1Fee(fee, msg, db, cfg)
+	}
+}
+
+// MakeSpecularEVMPreTransferHook creates specular's vm.EVMReader function
+// which is injected into the EVM and can be used to return the L1Fee of a transaction.
+// This is a read only method and does not change the state.
+func MakeSpecularL1FeeReader(cfg RollupConfig) vm.EVMReader {
+	log.Info("Injected Specular EVM reader")
+	log.Info("L1Oracle config", "address", cfg.GetL1OracleAddress(), "baseFeeSlot", cfg.GetL1OracleBaseFeeSlot())
+	return func(tx *types.Transaction, db vm.StateDB) (*big.Int, error) {
+		return calculateL1Fee(tx, db, cfg)
 	}
 }
 
@@ -88,7 +100,7 @@ func transactionFromMessage(msg types.Message, cfg RollupConfig) *types.Transact
 // L1Fee = L1FeeMultiplier * L1BaseFee * (TxDataGas + L1OverheadGas)
 // L1BaseFee is dynamically set by the L1Oracle
 // L1FeeMultiplier & L1OverheadGas are set in the rollup configuration
-func calculateL1Fee(tx *types.Transaction, evm *vm.EVM, cfg RollupConfig) (*big.Int, error) {
+func calculateL1Fee(tx *types.Transaction, db vm.StateDB, cfg RollupConfig) (*big.Int, error) {
 	// calculate L1 gas from RLP encoding
 	buf := new(bytes.Buffer)
 	if err := tx.EncodeRLP(buf); err != nil {
@@ -106,7 +118,7 @@ func calculateL1Fee(tx *types.Transaction, evm *vm.EVM, cfg RollupConfig) (*big.
 		zeroes, ones = zeroesAndOnes(rlp)
 
 		txDataGas    = big.NewInt(zeroes*txDataZero + (ones+txSignatureOverhead)*txDataOne + cfg.GetL1FeeOverhead())
-		basefee      = readStorageSlot(evm, cfg.GetL1OracleAddress(), cfg.GetL1OracleBaseFeeSlot())
+		basefee      = readStorageSlot(db, cfg.GetL1OracleAddress(), cfg.GetL1OracleBaseFeeSlot())
 		feeMutiplier = cfg.GetL1FeeMultiplier()
 
 		l1Fee       = new(big.Int).Mul(txDataGas, basefee)
@@ -145,22 +157,23 @@ func ScaleBigInt(num *big.Int, scalar float64) *big.Int {
 
 // subtract the L1 Fee from the sender of the Tx
 // add the Fee to the balance of the coinbase address
-func chargeL1Fee(l1Fee *big.Int, msg types.Message, evm *vm.EVM, cfg RollupConfig) error {
-	senderBalance := evm.StateDB.GetBalance(msg.From())
+func chargeL1Fee(l1Fee *big.Int, msg types.Message, db vm.StateDB, cfg RollupConfig) error {
+	senderBalance := db.GetBalance(msg.From())
+
 	if senderBalance.Cmp(l1Fee) < 0 {
 		return errors.New("insufficient balance to cover L1 fee")
 	}
 
-	evm.StateDB.AddBalance(cfg.GetCoinbase(), l1Fee)
-	evm.StateDB.SubBalance(msg.From(), l1Fee)
+	db.AddBalance(cfg.GetCoinbase(), l1Fee)
+	db.SubBalance(msg.From(), l1Fee)
 
 	log.Info("charged L1 Fee", "fee", l1Fee.Uint64())
 	return nil
 }
 
 // read the value from a given address / storage slot
-func readStorageSlot(evm *vm.EVM, address common.Address, slot common.Hash) *big.Int {
-	return evm.StateDB.GetState(address, slot).Big()
+func readStorageSlot(db vm.StateDB, address common.Address, slot common.Hash) *big.Int {
+	return db.GetState(address, slot).Big()
 }
 
 // zeroesAndOnes counts the number of 0 bytes and non 0 bytes in a byte slice
