+++ services/el_clients/geth/specular/cmd/geth/config.go
@@ -42,11 +42,11 @@ import (
 	"github.com/ethereum/go-ethereum/node"
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/naoina/toml"
-	specularUtils "github.com/specularl2/specular/clients/geth/specular/cmd/utils"
-	entry "github.com/specularl2/specular/clients/geth/specular/entry/geth"
-	"github.com/specularl2/specular/clients/geth/specular/internal/ethapi"
-	"github.com/specularl2/specular/clients/geth/specular/internal/flags"
-	"github.com/specularl2/specular/clients/geth/specular/rollup"
+	"github.com/specularl2/specular/lib/el_golang_lib/hook"
+	specularUtils "github.com/specularl2/specular/services/el_clients/geth/specular/cmd/utils"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/internal/ethapi"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/internal/flags"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup"
 )
 
 var (
@@ -196,7 +196,8 @@ func makeFullNode(ctx *cli.Context) (*node.Node, ethapi.Backend) {
 		rollupConfig := specularUtils.MakeRollupConfig(ctx, &cfg.Eth)
 
 		vmConfig := eth.BlockChain().GetVMConfig()
-		vmConfig.SpecularEVMPreTransferHook = entry.MakeSpecularEVMPreTransferHook(rollupConfig)
+		vmConfig.SpecularEVMPreTransferHook = hook.MakeSpecularEVMPreTransferHook(rollupConfig)
+		vmConfig.SpecularL1FeeReader = hook.MakeSpecularL1FeeReader(rollupConfig)
 
 		rollup.RegisterRollupService(stack, eth, eth.APIBackend, rollupConfig)
 	}
